+++ swift_browser_ui_frontend/src/components/ProgressBar.vue
@@ -1,14 +1,16 @@
 <template>
-  <div class="progress-bar">
+  <div
+    v-if="!finished"
+    class="progress-bar"
+  >
     <c-progress-bar
-      v-if="isUploading && progress != undefined"
+      v-if="progress != undefined"
       :value="(progress * 100).toFixed()"
       single-line
       :label="$t('message.upload.progressLabel')"
     />
-
     <c-progress-bar
-      v-else-if="isUploading"
+      v-else
       hide-details
     />
   </div>
@@ -17,21 +19,19 @@
 <script>
 export default {
   name: "ProgressBar",
+  props: ["type", "finished"],
   computed: {
-    isUploading() {
-      return this.$store.state.isUploading;
-    },
     progress() {
-      return this.$store.state.uploadProgress;
+      return this.type === "upload"
+        ? this.$store.state.uploadProgress
+        : this.$store.state.downloadProgress;
     },
   },
 };
 </script>
 
 <style scoped lang="scss">
-
 .progress-bar {
-  flex: 1
+  flex: 1;
 }
-
 </style>